:white_check_mark:Arranging the tests and adding Try/Catch
