+++ src/slack_api.py
@@ -1,5 +1,6 @@
 import os
 import requests
+from requests.exceptions import RequestException
 
 SLACK_API_URL = os.getenv("SLACK_API_URL")
 SLACK_ADMIN_INVITE_URL = os.getenv("SLACK_ADMIN_INVITE_URL")
@@ -8,8 +9,11 @@ SLACK_TOKEN = os.getenv("SLACK_TOKEN")
 
 def fetch_slack_users(token=None):
     headers = {"Authorization": f"Bearer {token or SLACK_TOKEN}"}
-    response = requests.get(SLACK_API_URL, headers=headers, verify=False)
-    return response.json()
+    try:
+        response = requests.get(SLACK_API_URL, headers=headers, verify=False)
+        return response.json()
+    except requests.RequestException as error:
+        return {"error": str(error)}
 
 
 def send_slack_invite(payload, token):
@@ -17,11 +21,14 @@ def send_slack_invite(payload, token):
         "Authorization": f"Bearer {token}",
         "Content-Type": "application/json",
     }
-    response = requests.post(
-        SLACK_ADMIN_INVITE_URL,
-        headers=headers,
-        json=payload,
-        timeout=10,
-        verify=False,
-    )
-    return response.json()
+    try:
+        response = requests.post(
+            SLACK_ADMIN_INVITE_URL,
+            headers=headers,
+            json=payload,
+            timeout=10,
+            verify=False,
+        )
+        return response.json()
+    except RequestException as error:
+        return {"error": str(error)}
